+++ src/main/java/com/market/place/controllers/BasketController.java
@@ -1,8 +1,10 @@
 package com.market.place.controllers;
 
 import com.market.place.models.Basket;
+import com.market.place.models.Product;
 import com.market.place.models.User;
 import com.market.place.services.impl.BasketServiceImpl;
+import com.market.place.services.impl.ProductServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.stereotype.Controller;
@@ -20,7 +22,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BasketController {
     private final BasketServiceImpl basketService;
-
+    private final ProductServiceImpl productService;
 
     @GetMapping("/baskets")
     String baskets(Model model, Principal principal){
@@ -39,8 +41,20 @@ public class BasketController {
         return "baskets/basket";
     }
 
-    @PostMapping("/add_product_in_basket")
-    String addProductInBasket(Long productId,Long basketId){
+    @GetMapping(value = "/select_basket/{productId}")
+    String addProductInBasket(@PathVariable Long productId, Model model,Principal principal){
+        User user = ((User) ((UsernamePasswordAuthenticationToken) principal).getPrincipal());
+
+        List<Basket> baskets = basketService.getAllByUserId(user.getId());
+        Product product = productService.getById(productId);
+
+        model.addAttribute("baskets",baskets);
+        model.addAttribute("product",product);
+        return "baskets/select_basket";
+    }
+
+    @GetMapping("/add_product_in_basket/{productId}/{basketId}")
+    String addProductInBasket(@PathVariable Long productId,@PathVariable Long basketId){
         basketService.addProduct(productId,basketId);
         return "redirect:/baskets/basket/" + basketId;
     }
