+++ src/main/java/com/market/place/controllers/ProductsController.java
@@ -2,7 +2,7 @@ package com.market.place.controllers;
 
 import com.market.place.models.ProductCategory;
 import com.market.place.models.ProductSubCategory;
-import com.market.place.services.impl.ProductServiceImpl;
+import com.market.place.services.impl.ProductCategoryServiceImpl;
 import com.market.place.services.impl.ProductSubCategoryServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -13,15 +13,35 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.security.Principal;
+import java.util.List;
 
 @Controller
-@RequestMapping("/products")
+@RequestMapping("menu/products")
 @RequiredArgsConstructor
 public class ProductsController {
+    private final ProductCategoryServiceImpl productCategoryService;
     private final ProductSubCategoryServiceImpl productSubCategoryService;
-    private final ProductServiceImpl productService;
+    @GetMapping("/product_category")
+    public String getAllCategory(Model model, Principal principal){
+        String role = ((UsernamePasswordAuthenticationToken) principal).getAuthorities().stream().toList().get(0).getAuthority();
 
-    @GetMapping(value = "/{productSubCategoryId}")
+        List<ProductCategory> productCategories = productCategoryService.getAll();
+        model.addAttribute("productCategories",productCategories);
+        model.addAttribute("userRole",role);
+        return "products/product_category";
+    }
+    @GetMapping(value = "/product_category/product_subcategory/{productCategoryId}")
+    public String getAllSubCategories(@PathVariable Long productCategoryId,Model model, Principal principal){
+        String role = ((UsernamePasswordAuthenticationToken) principal).getAuthorities().stream().toList().get(0).getAuthority();
+        ProductCategory productCategory = productCategoryService.getById(productCategoryId);
+
+        model.addAttribute("productSubCategories",productCategory.getProductSubCategories());
+        model.addAttribute("productCategory",productCategory);
+        model.addAttribute("userRole",role);
+
+        return "products/productSubCategory";
+    }
+    @GetMapping(value = "/product_category/product_subcategory/products/{productSubCategoryId}")
     public String getAllProducts(@PathVariable Long productSubCategoryId, Model model, Principal principal){  // изм назв
         String role = ((UsernamePasswordAuthenticationToken) principal).getAuthorities().stream().toList().get(0).getAuthority();
         ProductSubCategory productSubCategory = productSubCategoryService.getById(productSubCategoryId);
@@ -32,4 +52,6 @@ public class ProductsController {
 
         return "products/products";
     }
+
+
 }
