+++ src/main/java/com/market/place/repositories/BasketRepository.java
@@ -8,9 +8,6 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
 public interface BasketRepository extends JpaRepository<Basket,Long> {
-    @Query(value = "SELECT * FROM public.product_sub_category where product_category_id = :productCategoryId",nativeQuery = true)
-    List<ProductSubCategory> findAllByProductCategoryId(Long productCategoryId);
-
-    @Query(value = "SELECT * FROM public.basket where user_id = :userId",nativeQuery = true) // исп basket на baskets
+    @Query(value = "SELECT * FROM public.baskets where user_id = :userId",nativeQuery = true)
     List<Basket> findAllByUserId(Long userId);
 }
